fix: refactor code for fixing (clean code)
